# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BizEngine API
  version: 1.0.0
  description: Business engine for managing organizations, products, orders, warehouses, HR, finance, logistics, and Russian government integrations.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.bizengine.app
    description: Production
...
paths:
...
  /api/v1/organizations/{orgID}/messenger/conversations:
    get:
      tags: [Messenger]
      summary: List conversations
      operationId: listConversations
      parameters:
        - $ref: '#/components/parameters/orgID'
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Conversation list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        title: { type: string }
                        type: { type: string, enum: [direct, group, entity] }
                        entity_type: { type: string }
                        entity_id: { type: string, format: uuid }
                        last_message_at: { type: string, format: date-time }
                        unread_count: { type: integer }
                        muted: { type: boolean }
    post:
      tags: [Messenger]
      summary: Create conversation
      operationId: createConversation
      parameters:
        - $ref: '#/components/parameters/orgID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [data]
              properties:
                data:
                  type: object
                  required: [type, member_ids]
                  properties:
                    title: { type: string }
                    type: { type: string, enum: [direct, group] }
                    member_ids: { type: array, items: { type: string, format: uuid } }
      responses:
        '201':
          description: Conversation created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      title: { type: string }
                      type: { type: string }
                      created_at: { type: string, format: date-time }
  /api/v1/organizations/{orgID}/messenger/conversations/{convID}:
    get:
      tags: [Messenger]
      summary: Get conversation details
      operationId: getConversation
      parameters:
        - $ref: '#/components/parameters/orgID'
        - name: convID
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Conversation details
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      title: { type: string }
                      type: { type: string }
                      entity_type: { type: string }
                      entity_id: { type: string, format: uuid }
                      members:
                        type: array
                        items:
                          type: object
                          properties:
                            user_id: { type: string, format: uuid }
                            name: { type: string }
                            role: { type: string }
                      created_at: { type: string, format: date-time }
...
  /api/v1/organizations/{orgID}/messenger/conversations/{convID}/mute:
    post:
      tags: [Messenger]
      summary: Mute or unmute conversation
      operationId: muteConversation
      parameters:
        - $ref: '#/components/parameters/orgID'
        - name: convID
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [muted]
              properties:
                muted: { type: boolean }
      responses:
        '200':
          description: Mute state updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
  /api/v1/organizations/{orgID}/messenger/conversations/{convID}/messages:
    get:
      tags: [Messenger]
      summary: List messages in conversation
      operationId: listMessages
      parameters:
        - $ref: '#/components/parameters/orgID'
        - name: convID
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: before
          in: query
          description: Cursor — message ID to paginate before
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Message list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        conversation_id: { type: string, format: uuid }
                        sender_id: { type: string, format: uuid }
                        content: { type: string }
                        content_type: { type: string, enum: [text, system, file] }
                        metadata: { type: object }
                        created_at: { type: string, format: date-time }
                        updated_at: { type: string, format: date-time }
                        reactions:
                          type: array
                          items:
                            type: object
                            properties:
                              emoji: { type: string }
                              user_ids: { type: array, items: { type: string, format: uuid } }
    post:
      tags: [Messenger]
      summary: Send message
      operationId: sendMessage
      parameters:
        - $ref: '#/components/parameters/orgID'
        - name: convID
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
                content_type: { type: string, enum: [text, file], default: text }
                metadata: { type: object }
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      conversation_id: { type: string, format: uuid }
                      sender_id: { type: string, format: uuid }
                      content: { type: string }
                      content_type: { type: string }
                      created_at: { type: string, format: date-time }
  /api/v1/organizations/{orgID}/messenger/conversations/{convID}/messages/{msgID}:
    put:
      tags: [Messenger]
      summary: Edit message
      operationId: editMessage
      parameters:
        - $ref: '#/components/parameters/orgID'
        - name: convID
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: msgID
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
      responses:
        '200':
          description: Message updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      content: { type: string }
                      updated_at: { type: string, format: date-time }
    delete:
      tags: [Messenger]
      summary: Delete message
      operationId: deleteMessage
      parameters:
        - $ref: '#/components/parameters/orgID'
        - name: convID
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: msgID
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Message deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
...
  /api/v1/organizations/{orgID}/messenger/unread-count:
    get:
      tags: [Messenger]
      summary: Get total unread message count
      operationId: getUnreadCount
      parameters:
        - $ref: '#/components/parameters/orgID'
      responses:
        '200':
          description: Unread count
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      total: { type: integer }
  /api/v1/organizations/{orgID}/messenger/messages/{msgID}/reactions:
    post:
      tags: [Messenger]
      summary: Add reaction to message
      operationId: addReaction
      parameters:
        - $ref: '#/components/parameters/orgID'
        - name: msgID
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [emoji]
              properties:
                emoji: { type: string }
      responses:
        '200':
          description: Reaction added
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
  /api/v1/organizations/{orgID}/messenger/messages/{msgID}/reactions/{emoji}:
    delete:
      tags: [Messenger]
      summary: Remove reaction from message
      operationId: removeReaction
      parameters:
        - $ref: '#/components/parameters/orgID'
        - name: msgID
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: emoji
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Reaction removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
...
  /api/v1/organizations/{orgID}/messenger/search:
    get:
      tags: [Messenger]
      summary: Search messages
      operationId: searchMessages
      parameters:
        - $ref: '#/components/parameters/orgID'
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: conversation_id
          in: query
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        conversation_id: { type: string, format: uuid }
                        sender_id: { type: string, format: uuid }
                        content: { type: string }
                        content_type: { type: string }
                        created_at: { type: string, format: date-time }
components:
  parameters:
    orgID:
      name: orgID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...